requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
langchain-chroma
langchain-huggingface
chromadb
sentence-transformers
google-genai
pandas
numpy
pypdf
